% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{translate}
\alias{translate}
\title{Translate Sentences}
\usage{
translate(text, target_lang)
}
\arguments{
\item{text}{String to be translated.}

\item{target_lang}{Language code of language user wishes to translate sentence to.
Current options are: "BG”  (Bulgarian), "CS" (Czech), "DA" (Danish), "DE" (German),
"EL" (Greek), "EN-GB" (British English), "EN-US" (American English), "ES" (Spanish),
"ET" (Estonian), "FI" (Finnish), "FR" (French), "HU" (Hungarian), "ID" (Indonesian),
"IT" (Italian), "JA" (Japanese), "LT" (Lithuanian), "LV" (Latvian), "NL" (Dutch),
"PL" (Polish), "PT-PT" (Portuguese), "PT-BR" (Brazilian Portuguese), "RO" (Romanian),
"RU" (Russian), "SK" (Slovak), "SL" (Slovenian), "SV" (Swedish), "TR" (Turkish), and
"ZH" (Chinese)}
}
\value{
The translated version of \code{text} in target language \code{target_lang}
}
\description{
Translates text into a target language. Original language does
not need to be specified as function automatically recognises it.
}
